Fixed initialization bug and changed test script

+++ SimulationLib/IncidenceTimeSeries-inl.h
@@ -97,7 +97,7 @@ namespace SimulationLib {
         //    Record was called. In this case, we pass the period number, rather
         //    than the time.
         if (stats && recordPeriod == RECORD_ON_ALL)
-            stats->Record(time, (double)aggregatedObservation + value);
+            stats->Record(time, (double)aggregatedObservation + (double)value);
         else if (stats &&
                  currentPeriod > lastPeriod &&
                  (lastPeriod % recordPeriod) == 0)
